+++ .secrets.baseline
@@ -1097,5 +1097,5 @@
       }
     ]
   },
-  "generated_at": "2026-07-01T10:49:14Z"
+  "generated_at": "2026-07-01T10:57:50Z"
 }
