+++ src/test/java/uk/gov/pay/connector/gatewayaccount/service/AdyenAccountSetupServiceTest.java
@@ -29,7 +29,7 @@ import static uk.gov.pay.connector.gatewayaccount.model.GatewayAccountType.TEST;
 import static uk.gov.pay.connector.gatewayaccountcredentials.model.GatewayAccountCredentialsEntityFixture.aGatewayAccountCredentialsEntity;
 
 @ExtendWith(MockitoExtension.class)
-class AdyenAccountSetupResponseServiceTest {
+class AdyenAccountSetupServiceTest {
 
     @Mock
     private AdyenAccountSetupDao mockAdyenAccountSetupDao;
